PgsToSrtWrapper/requirements.txt
ffmpeg-python==0.2.0
future==0.18.2
pathtools==0.1.2
